pyproject.toml
[tool.poetry]
authors = ["Bartosz Nowakowski <b.nowakowski112@gmail.com>"]
description = ""
name = "wykladowca-pl-v2"
packages = [{include = "wykladowca"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
django = "~4.2.2"
django-quill-editor = "^0.1.40"
pillow = "^9.5.0"
pytest = "^7.3.2"
pytest-django = "^4.5.2"
python = "^3.9"
python-dotenv = "^1.0.0"
zeep = "~4.2.1"
djangorestframework = "~3.14.0"
requests = "^2.31.0"
lxml = "~4.9.2"
celery = {version = "~5.3.0", extras = ["redis"]}
django-celery-results = "^2.5.1"
html2text = "^2020.1.16"
pydantic = "^1.10.9"
flower = "^2.0.0"
dnspython = "^2.3.0"
borb = "^2.1.15"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8-django = "^1.3"
pylint = "^2.17.4"
pylint-django = "^2.5.3"
ipython = "^8.14.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
